requirements.txt
streamlit
pandas
python-docx
langchain-openai
langchain-core
openpyxl